requirements.txt
pandas==2.1.4 
sqlalchemy==2.0.23 
psycopg2-binary==2.9.9 
faker==25.0.0 
pyyaml==6.0.1 
pytest==8.0.0 
pytest-cov==5.0.0 
